+++ animation_nodes/ui/node_menu.py
@@ -519,6 +519,7 @@ class SplineMenu(bpy.types.Menu):
         insertNode(layout, "an_TransformSplineNode", "Transform")
         insertNode(layout, "an_ConnectSplinesNode", "Connect")
         insertNode(layout, "an_TrimSplineNode", "Trim")
+        insertNode(layout, "an_BevelSplineNode", "Bevel")
         insertNode(layout, "an_SetSplineRadiusNode", "Set Radius")
         insertNode(layout, "an_MakeSplineCyclicNode", "Make Cyclic")
         insertNode(layout, "an_SmoothBezierSplineNode", "Smooth Bezier")